fix(test): Split router into own module
