+++ apps/editor.planx.uk/src/index.tsx
@@ -7,23 +7,19 @@ import CssBaseline from "@mui/material/CssBaseline";
 import { StyledEngineProvider, ThemeProvider } from "@mui/material/styles";
 import { MyMap } from "@opensystemslab/map";
 import { QueryClientProvider } from "@tanstack/react-query";
-import {
-  createRouter,
-  RouterProvider as TanStackRouterProvider,
-} from "@tanstack/react-router";
+import { RouterProvider as TanStackRouterProvider } from "@tanstack/react-router";
 import { ToastContextProvider } from "contexts/ToastContext";
-import { getCookie, setCookie } from "lib/cookie";
 import { initFeatureFlags } from "lib/featureFlags";
 import { queryClient } from "lib/queryClient";
 import { AnalyticsProvider } from "pages/FlowEditor/lib/analytics/provider";
 import React, { Suspense, useEffect } from "react";
 import { createRoot } from "react-dom/client";
 import { ToastContainer } from "react-toastify";
+import { router } from "router";
 
 // init airbrake before everything else
 import * as airbrake from "./airbrake";
 import { client } from "./lib/graphql";
-import { routeTree } from "./routeTree.gen";
 import { defaultTheme } from "./theme";
 
 if (import.meta.env.VITE_APP_ENV !== "production") {
@@ -45,41 +41,6 @@ window.addEventListener("vite:preloadError", (event) => {
 });
 
 initFeatureFlags();
-
-const hasJWT = (): boolean | void => {
-  // This cookie indicates the presence of the secure httpOnly "jwt" cookie
-  const authCookie = getCookie("auth");
-  if (authCookie) return true;
-
-  // If JWT not set via cookie, check search params
-  const jwtSearchParams = new URLSearchParams(window.location.search).get(
-    "jwt",
-  );
-  if (!jwtSearchParams) return false;
-
-  // Remove JWT from URL, and re-run this function
-  setCookie("jwt", jwtSearchParams);
-  setCookie("auth", JSON.stringify({ loggedIn: true }));
-  // Remove the jwt param from the URL
-  const url = new URL(window.location.href);
-  url.searchParams.delete("jwt");
-  window.history.replaceState({}, document.title, url.pathname + url.search);
-
-  // Return true to indicate authenticated state
-  return true;
-};
-
-export const router = createRouter({
-  routeTree,
-  context: { currentUser: hasJWT() },
-  scrollRestoration: true,
-});
-
-declare module "@tanstack/react-router" {
-  interface Register {
-    router: typeof router;
-  }
-}
 
 const Layout: React.FC<{
   children: React.ReactNode;
